[mono] Use OP_MOVE when converting from l to i
OP_SEXT_I4 does a sign extend from i4 to i8, meaning that it has an lreg as destination. The original OP_LCONV_TO_I4 has an ireg as destination. This resulted in a buffer overflow in cprop, since on 32bit, a lreg is comprised of 3 regs (2 additional for HI and LO), which were not allocated.