[llvm] Link LLVM frames together
Allocate frame, store the current parent into it, pass it as an argument to all further calls. Every frame contains a pointer to the start of the method. This is fast to implement in llvm and it enables us to find the method by looking up the jit info tables.