[mono][interp] Implement offset allocation for struct field vars
Their liveness is tied to the parent vt, their offset is not allocated during general offset allocation. When we use these vars, they extend the liveness of the parent struct. Once we finished allocating the offsets for normal vars, we compute their real offsets relative to the offset of the parent struct var.