[mono][interp] Enable copy propagation through struct field vars
Partial implementation for reference. The problem with this implementation is that it is annoying to deal with both vt and the promoted fields. Ideally we would be able to remove all references of the vt, in which case we would be left only with the promoted fields which no longer need to be tied to the VT (so they could be passed as an argument for example). In order to achieve this we would probably need SSA so we can easily delete the initialization of the struct. This is a long due optimization that we could use to clear up current optimization passes and allow for other future passes. (We would probably need to deal with non-ssa locals when ldloca is used, how exactly we would do this is unclear at this point)