[mono][interp] Fix catching of stack overflow exception
When unwinding during a stack overflow we ensure that sufficient stack has been freed before we attempt to catch the exception. If all the frames were coming from interpreter, the check would never be true because free_stack only counts the native stack and the interpreter runs in non-recursive mode, allocating the stack in a separate memory space. If we are unwinding interpreter frames, take into account also the interpreter stack space.