[mono][interp] Fix diverging var offsets between tiered and untiered methods
Patchpoints introduce some complications since some variables have to be
accessed from same offset between unoptimized and optimized methods.

Consider the following scenario
BB0 -> BB2       BB0: TMP <- def; IL_VAR <- TMP
| ^ _            BB1: Use IL_VAR
v | /|           BB2: Use IL_VAR
BB1

BB1 is a basic block containing a patchpoint, BB0 dominates both BB1 and BB2.
IL_VAR is used both in BB1 and BB2. In BB1, in optimized code, we could normally
replace use of IL_VAR with use of TMP. However, this is incorrect, because TMP
can be allocated at a different offset from IL_VAR and, if we enter the method
from the patchpoint in BB1, the data at var TMP would not be initialized since
we only copy the IL var space.

Even if we prevent the copy propagation in BB1, then tiering is still broken.
In BB2 we could replace use of IL_VAR with TMP, and we end up hitting the same problem.
Optimized code will attempt to access value of IL_VAR from the offset of TMP_VAR,
which is not initialized if we enter from the patchpoint in BB1.

We solve these issues by inserting a MINT_DEF_TIER_VAR in BB1. This instruction
prevents cprop of the IL_VAR in the patchpoint bblock since MINT_DEF_TIER_VAR is seen
as a redefinition. In addition to that, in BB2 we now have 2 reaching definitions for
IL_VAR, the original one from BB0 and the one from patchpoint bblock from BB1. This
will force a phi definition in BB2 and we will once again be forced to access IL_VAR
from the original offset that is equal to the one in unoptimized method.