[mono][interp] Properly implement Unsafe.SkipInit
Before this change, we were simply ignoring it, generating a NOP. We now actually emit a MINT_DEF opcode. This opcode has a dreg so it will serve as an initializer for the var, but only at IR level, since no actual instruction will be generated. This will allow to kill any previous, unused, initializers of the var.