[mono][aot] Improve aot-skip debug flag
Back in the mono/mono repo we used to have a simple test that we would run with different values for aot-skip. This parameter would ignore loading the n-th method from the aot images. This is not very useful for stress testing aot-interp transitions. Instead we pass two values to the option and we use these values to select a subset of methods that we are not loading from the aot images.