Add support for newobj
The newobj instruction receives the set of arguments and it needs to create the object in question, add it as the `this` argument and return it as well. All this work is done by a single interpreter newobj instruction. Once handling of the new object, the same call path is reused for executing the constructor. The newobj instruction thinks it receives the `this` pointer as an argument, so we can reuse the offset allocator to generate the exact offsets for each argument. We dummy initialize `this` via an intermediate opcode, so the offset allocator handles this pattern cleanly.